Fix menu options
Fix menu options
