+++ src/routes/[network]/(account)/+layout.svelte
@@ -1,34 +1,47 @@
 <script lang="ts">
-	import AccountNavigation from '$lib/components/navigation/accountnavigation.svelte';
+	import type { UnicoveContext } from '$lib/state/client.svelte';
 	import { page } from '$app/stores';
+	import { getContext } from 'svelte';
+	import { Stack } from '$lib/components/layout';
+	import PillGroup from '$lib/components/navigation/pillgroup.svelte';
 
-	const { children, data } = $props();
+	const { children } = $props();
+	const context = getContext<UnicoveContext>('state');
 
-	const destinations = $derived.by(() => {
-		const network = String(data.network);
+	const tabOptions = $derived.by(() => {
+		const account = String(context.account?.name);
+		const network = String(context.account?.network);
 		return [
-			{ href: `/${network}/account`, text: 'Account' },
-			{ href: `/${network}/move`, text: 'Move' },
-			{ href: `/${network}/permissions`, text: 'Permissions' },
-			{ href: `/${network}/ram`, text: 'RAM' },
-			{ href: `/${network}/resources`, text: 'Resources' },
-			{ href: `/${network}/send`, text: 'Send' },
-			{ href: `/${network}/transactions`, text: 'Transactions' },
-			{ href: `/${network}/vote`, text: 'Vote' }
+			{ href: `/${network}/account/${account}`, text: 'Overview' },
+			{ href: `/${network}/account/${account}/activity`, text: 'Activity' },
+			{ href: `/${network}/account/${account}/balances`, text: 'Balances' },
+			{ href: `/${network}/account/${account}/permissions`, text: 'Permissions' },
+			{ href: `/${network}/account/${account}/ram`, text: 'RAM' },
+			{ href: `/${network}/account/${account}/resources`, text: 'Resources' },
+			{ href: `/${network}/account/${account}/staked`, text: 'Staked' },
+			{ href: `/${network}/account/${account}/chaindata`, text: 'Chain Data' }
 		];
 	});
 
-	let rootPathname = $derived($page.url.pathname.split('/').slice(2)[1]);
+	let currentTab = $derived($page.url.pathname.split('/').slice(2)[3]);
 
 	// Derive the active state of each destination
 	let options = $derived(
-		destinations.map((destination) => ({
-			...destination,
-			active: destination.href.includes(rootPathname)
+		tabOptions.map((option) => ({
+			...option,
+			active: option.href.split('/').slice(2)[2] === currentTab
 		}))
 	);
 </script>
 
-<AccountNavigation {options} />
+<header class="layout-stack gap-6">
+	<Stack class="gap-2">
+		<h1 class="h2 leading-none text-white/60">Account</h1>
+		{#if context.account}
+			<h2 class="h1 font-bold text-white">{context.account.name}</h2>
+		{/if}
+	</Stack>
+	<PillGroup {options} class="mb-6" />
+</header>
 
 {@render children()}
